refactor: renamed all components and requierements for PascalCase
